pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thinkingproducts"
version = "0.1.0"
description = "AI agents for data querying, Python analysis, and chart generation."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
authors = [
    { name = "thinking products"},
]
keywords = ["ai", "agents", "data-analysis", "charts", "llm", "gemini", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "deepagents>=0.3.5",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.0",
    "httpx>=0.27.2",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "mcp>=1.25.0",
    "numpy>=2.4.1",
    "psycopg[binary]>=3.1.18",
    "psycopg2-binary>=2.9.9",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/thinkingproducts/thinkingproducts"
Documentation = "https://github.com/thinkingproducts/thinkingproducts#readme"
Repository = "https://github.com/thinkingproducts/thinkingproducts"
Issues = "https://github.com/thinkingproducts/thinkingproducts/issues"

[project.scripts]
tp = "thinkingproducts.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["thinkingproducts"]

[tool.hatch.build.targets.sdist]
include = [
    "thinkingproducts/**/*.py",
    "thinkingproducts/**/*.md",
    "thinkingproducts/mcps/Dockerfile.*",
    "thinkingproducts/mcps/resources/*",
    "docker-compose.yml",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short", "-ra"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
